Starting the refactor Translator Registry access code (baby steps...)
